# lib/api-zod/src/generated/types/createDepartmentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DS Engineosys API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDepartmentRequest {
  name: string;
  parentId?: string | null;
  description?: string | null;
}