# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DS Engineosys API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileRole } from "./userProfileRole";

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  role: UserProfileRole;
  createdAt: Date;
}